# order-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: Manages users and delivery orders
  version: 1.0.0

servers:
  - url: http://localhost:8090

tags:
  - name: users
    description: User registration and login
  - name: orders
    description: Order management

paths:

  /api/users/register:
    post:
      tags: [users]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
        '409':
          description: Username already taken

  /api/users/login:
    post:
      tags: [users]
      summary: Login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
        '401':
          description: Invalid credentials

  /api/orders:
    post:
      tags: [orders]
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId, pickupStreet, pickupLat, pickupLng, deliveryStreet, deliveryLat, deliveryLng, weightKg]
              properties:
                customerId:
                  type: string
                pickupStreet:
                  type: string
                pickupLat:
                  type: number
                pickupLng:
                  type: number
                deliveryStreet:
                  type: string
                deliveryLat:
                  type: number
                deliveryLng:
                  type: number
                weightKg:
                  type: number
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid input

  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Get order details
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /api/orders/{id}/confirm:
    post:
      tags: [orders]
      summary: Confirm a pending order
      operationId: confirmOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be confirmed

  /api/orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Cancel a pending order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be cancelled

  /api/orders/{id}/complete:
    post:
      tags: [orders]
      summary: Mark order as completed
      operationId: completeOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order completed
        '400':
          description: Order cannot be completed

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        pickupStreet:
          type: string
        deliveryStreet:
          type: string
        weightKg:
          type: number
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        createdAt:
          type: string